Refactor type specs
